flask/requirements.txt
Flask
google-generativeai
requests
pymongo
flask_pymongo
pandas
unidecode
python-dotenv